% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_Rdata.r
\name{load_all_Rdata}
\alias{load_all_Rdata}
\title{Load all .Rdata or .rdata files in a directory}
\usage{
load_all_Rdata(directory)
}
\arguments{
\item{directory}{A character string of the directory path to scan for .Rdata or .rdata files.}
}
\description{
This function scans a specified directory for all files with .Rdata or .rdata
extensions and loads them into the R environment. It provides error messages
for non-existent directories and directories with no relevant files.
}
